+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ### Features:
 - Added !coords party command to send your current coordinates (Aliases: !c, !xyz)
 - Added !ptme alias to the !pt party command
+- Improved !since command:
+  - Added support for rare drops and dyes.
+  - Added optional <username> parameter to filter who should respond.
+  - Now displays related rare drops when checking a sea creature.
 
 ### Changes: 
 -